- add output domain args to fused stencil programs - add option to switch from single fused stencil call in diffusion to gt4py program - load grid indices as int (instead of int32)

+++ atm_dyn_iconam/src/icon4py/diffusion/icon_grid.py
@@ -102,8 +102,8 @@ class IconGrid:
     def with_start_end_indices(
         self, dim: Dimension, start_indices: np.ndarray, end_indices: np.ndarray
     ):
-        self.start_indices[dim] = start_indices
-        self.end_indices[dim] = end_indices
+        self.start_indices[dim] = start_indices.astype(int)
+        self.end_indices[dim] = end_indices.astype(int)
 
     @builder
     def with_connectivities(self, connectivity: Dict[Dimension, np.ndarray]):
@@ -130,7 +130,7 @@ class IconGrid:
 
     def get_indices_from_to(
         self, dim: Dimension, start_marker: int, end_marker: int
-    ) -> Tuple[int32, int32]:
+    ) -> Tuple[int, int]:
         """
         Use to specifzy domains of a field for field_operator.
